requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2020.6.24=0
certifi=2020.6.20=py37_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
ncurses=6.2=he6710b0_1
openssl=1.1.1g=h7b6447c_0
pip=20.2.1=py37_0
pycups=2.0.1=pypi_0
pypdf2=1.26.0=pypi_0
python=3.7.7=hcff3b4d_5
readline=8.0=h7b6447c_0
setuptools=49.2.1=py37_0
sqlite=3.32.3=h62c20be_0
tk=8.6.10=hbc83047_0
wheel=0.34.2=py37_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
